* Tests fixed

+++ src/main/kotlin/learn/coroutine/domain/task/SequentialTaskUseCase.kt
@@ -26,6 +26,6 @@ class SequentialTaskUseCase(
 
         delayTask(taskDuration)
 
-        return@backgroundTask TaskExecutionSuccess(fetchedData)
+        return@backgroundTask TaskExecutionSuccess(fetchedData.toDouble())
     }
 }
