+++ src/main/kotlin/learn/coroutine/model/TaskExecutionResult.kt
@@ -2,6 +2,6 @@ package learn.coroutine.model
 
 
 sealed class TaskExecutionResult
-data class TaskExecutionSuccess(val result: Long) : TaskExecutionResult()
+data class TaskExecutionSuccess(val result: Double) : TaskExecutionResult()
 //data class TaskExecutionError(val exception: CustomTaskException) : TaskExecutionResult()
 object TaskExecutionCancelled : TaskExecutionResult()
